requirements.txt
# 🧠 Core Frameworks
streamlit
flask
python-dotenv
gunicorn

# 🤖 LLM & Vector Store
groq
langchain-groq
langchain
langchain-community
langchain-core
astrapy
cassandra-driver

# 📄 Document Processing
pypdf
python-docx
python-pptx
unstructured[easy_paddle]  # For image and complex PDF parsing
pytesseract
opencv-python-headless
beautifulsoup4
lxml
html2text

# 🔍 Embeddings & Machine Learning
sentence-transformers
torch
torchvision
torchaudio
transformers
scikit-learn

# 🔧 Utilities
requests
numpy
pandas
pyarrow
fastapi
uvicorn



# Core Frameworks
flask
python-dotenv
gunicorn

# LLM & Vector Store
groq
langchain-groq
langchain
langchain-community
langchain-core
astrapy
cassandra-driver

# Document Processing
pypdf
python-docx
python-pptx
unstructured[easy_paddle]
pytesseract
opencv-python-headless
beautifulsoup4
lxml
html2text

# Embeddings & ML
sentence-transformers
torch
torchvision
torchaudio
transformers
scikit-learn

# Utilities
requests
numpy
pandas
pyarrow
fastapi
uvicorn


langchain-astradb # <-- New package for AstraDB
langchain-huggingface # <-- New package for embeddings
astrapy # The base library for Astra